chore [SRTP-819]: rename file to reflect test name
